+++ Server/DAL/Entities/Deposit.cs
@@ -13,6 +13,6 @@ namespace MoneyManager.DAL.Entities
 
         public short Percentage { get; set; }
 
-        public int InitialAmount { get; set; }
+        public decimal InitialAmount { get; set; }
     }
 }
